iot based food industry monitoring system.
project related to Developing an Internet of Things (IoT) based system to track fruits that are delivered to nearby vendors or to customers directly. To keep an eye on the location, the surrounding temperature and humidity, and the freshness of the fruits-in this case, the apple-using various gas and odor sensors. Moreover, the device will use image processing to determine if the apple is rotting